리팩토링 수행
