--- back/src/main/java/com/example/back/service/RouteTiimeService.java
@@ -1,8 +1,0 @@
-package com.example.back.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class RouteTiimeService {
-
-}
